Create 16 by 16 grid in center
After a painstaking few hours trying to figure out how to do the format
for the 16 by 16 grid, finally settled on using rows of containers that
contains the individual square divs, all of which are wrapped in a big
main container in body. I think it is possible to have just one large
container containing all the divs, with flex-wrap: wrap but requires
more formatting? Javascript code was also changed to facilitate the
creation of the rows of divs.