Color boxes when mouse hover over
Boxes can now be colored when the mouse hovers over them. For additional
aesthetic, the cursor will change to a pointer to affirm users that it
is intended and something is happening. Colors are also randomised but
I don't understand the other extra credit where you make the Colors
darker and darker till black as then it wouldn't feel like the color is
random.