Add button to change grid size and clean up code
Button created in the centre above the grid to prompt user for a number
to change the grid size when pressed. It will allow the previous grid to
be erased and create a new grid of a different size the user inputs.

Code has also been cleaned up to be more modular, especially the
Javascript file, to prevent repetition of code. New createGrid()
function made to fulfil the aforementioned purpose.